src/data/script/requirements.txt
python-dotenv>=1.0.0
requests>=2.25.0
yfinance>=0.2.0
pandas>=1.3.0
